Create Increasing Problems - Try out
Problem Statement
The solution we came up with has a key problem. Data for mobiles and shoes are mixed up in the same code. If we have to make changes to the logic for purchasing shoes, we have to modify method that has logic for both shoes and mobiles. For example, if we have to add 'material' to the shoe and have a 5% tax on 'leather' shoes, then we have to go through code related to mobile as well, as shown below: